+++ Libraries/Core/Models/Eft/Game/GetRaidTimeResponse.cs
@@ -5,13 +5,13 @@ namespace Core.Models.Eft.Game;
 public record GetRaidTimeResponse
 {
     [JsonPropertyName("RaidTimeMinutes")]
-    public int? RaidTimeMinutes { get; set; }
+    public double? RaidTimeMinutes { get; set; }
 
     [JsonPropertyName("NewSurviveTimeSeconds")]
-    public int? NewSurviveTimeSeconds { get; set; }
+    public double? NewSurviveTimeSeconds { get; set; }
 
     [JsonPropertyName("OriginalSurvivalTimeSeconds")]
-    public int? OriginalSurvivalTimeSeconds { get; set; }
+    public double? OriginalSurvivalTimeSeconds { get; set; }
 
     [JsonPropertyName("ExitChanges")]
     public List<ExtractChange>? ExitChanges { get; set; }
@@ -23,11 +23,11 @@ public record ExtractChange
     public string? Name { get; set; }
 
     [JsonPropertyName("MinTime")]
-    public int? MinTime { get; set; }
+    public double? MinTime { get; set; }
 
     [JsonPropertyName("MaxTime")]
-    public int? MaxTime { get; set; }
+    public double? MaxTime { get; set; }
 
     [JsonPropertyName("Chance")]
-    public int? Chance { get; set; }
+    public double? Chance { get; set; }
 }
